replace system components with auth-ui
