+++ lib/app/package.json
@@ -27,7 +27,7 @@
     "remark-gfm": "^4.0.0",
     "remark-math": "^6.0.0",
     "zustand": "^5.0.2",
-    "@ninembs-studio/auth-ui": "0.0.1-alpha.1"
+    "@ninembs-studio/auth-ui": "0.0.1-alpha.6"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
