+++ lib/app/src/app/dashboard/page.tsx
@@ -1,12 +1,13 @@
 "use client";
 
-import { AuthWrapper } from "@/components/system/AuthWrapper";
-import { AuthHeader } from "@/components/system/AuthHeader";
+import { useRouter } from "next/navigation";
+import { AuthWrapper, AuthHeader } from "@ninembs-studio/auth-ui";
 import { ChatHistory } from "@/components/ChatHistory";
 
 export default function Dashboard() {
+  const router = useRouter();
   return (
-    <AuthWrapper>
+    <AuthWrapper history={router}>
       <div className="max-w-6xl mx-auto p-4 h-full">
         <AuthHeader />
         <ChatHistory />
