+++ lib/app/src/app/register/page.tsx
@@ -1,9 +1,12 @@
-import Register from "@/components/system/Registration";
+import { Registration, PublicHeader } from "@ninembs-studio/auth-ui";
+import { useRouter } from "next/navigation";
 
 export default function RegisterPage() {
+  const router = useRouter();
   return (
     <div>
-      <Register />
+      <PublicHeader />
+      <Registration history={router} />
     </div>
   );
 }
