+++ lib/app/src/app/reset-password/page.tsx
@@ -1,11 +1,22 @@
-import { Suspense } from "react";
-import ResetPasswordClient from "./ResetPasswordClient";
+"use client";
 
-// this is the server component that wraps our client component in a suspense boundary
+import { useSearchParams, useRouter } from "next/navigation";
+import { Card } from "@/components/ui/Card";
+import { ResetPassword } from "@ninembs-studio/auth-ui";
+
 export default function ResetPasswordPage() {
-  return (
-    <Suspense fallback={<div className="text-white">loading...</div>}>
-      <ResetPasswordClient />
-    </Suspense>
-  );
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const token = searchParams.get("token");
+  const appId = searchParams.get("appId");
+
+  if (!token || !appId) {
+    return (
+      <Card className="p-6">
+        <p className="text-red-500">invalid reset password link</p>
+      </Card>
+    );
+  }
+
+  return <ResetPassword token={token} appId={appId} history={router} />;
 }
