--- lib/app/src/components/system/Readme.system.md
@@ -1,5 +1,0 @@
-# System Components
-
-These are core to the applications shell and include functionality needed to authorize and authenticate a user via a JWT as a refresh token. One day i'll move this to a package, but right now it's all just scoped to this directory
-
-it's also still a wip (when compared to the auth_v3 openapi spec)
