--- lib/app/src/components/system/settings/BillingSettings.tsx
@@ -1,12 +1,0 @@
-"use client";
-
-export function BillingSettings() {
-  return (
-    <>
-      <h2 className="card-title text-xl mb-6">Billing Settings</h2>
-      <p className="text-base-content/70">
-        Manage your subscription and billing information.
-      </p>
-    </>
-  );
-}
