+++ lib/auth_uisdk/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ninembs-studio/auth-ui",
-  "version": "0.0.1-alpha.1",
+  "version": "0.0.1-alpha.6",
   "private": false,
   "scripts": {
     "build": "tsc",
@@ -12,46 +12,7 @@
   "module": "dist/index.js",
   "types": "dist/index.d.ts",
   "exports": {
-    "./auth-header": {
-      "import": "./dist/components/AuthHeader.js",
-      "require": "./dist/components/AuthHeader.js",
-      "types": "./dist/components/AuthHeader.d.ts"
-    },
-    "./auth-wrapper": {
-      "import": "./dist/components/AuthWrapper.js",
-      "require": "./dist/components/AuthWrapper.js",
-      "types": "./dist/components/AuthWrapper.d.ts"
-    },
-    "./email-verification": {
-      "import": "./dist/components/EmailVerification.js",
-      "require": "./dist/components/EmailVerification.js",
-      "types": "./dist/components/EmailVerification.d.ts"
-    },
-    "./forgot-password": {
-      "import": "./dist/components/ForgotPassword.js",
-      "require": "./dist/components/ForgotPassword.js",
-      "types": "./dist/components/ForgotPassword.d.ts"
-    },
-    "./login": {
-      "import": "./dist/components/Login.js",
-      "require": "./dist/components/Login.js",
-      "types": "./dist/components/Login.d.ts"
-    },
-    "./logout-button": {
-      "import": "./dist/components/LogoutButton.js",
-      "require": "./dist/components/LogoutButton.js",
-      "types": "./dist/components/LogoutButton.d.ts"
-    },
-    "./registration": {
-      "import": "./dist/components/Registration.js",
-      "require": "./dist/components/Registration.js",
-      "types": "./dist/components/Registration.d.ts"
-    },
-    "./reset-password": {
-      "import": "./dist/components/ResetPassword.js",
-      "require": "./dist/components/ResetPassword.js",
-      "types": "./dist/components/ResetPassword.d.ts"
-    }
+    ".": "./dist/index.js"
   },
   "dependencies": {
     "@radix-ui/react-tooltip": "^1.1.6",
